feat: added an option for prettier and a extended eslint config (#1425)
* added import order sorting

* cli format

* format www

* .

* done?

* lint

* we do a little reverse

* fixed format:check writing

* smoll changes

* split out installers

* removed eslint from base project

* added eslint back

* formatting

* fixed linting

* fixed base project

* rip ci time

* Update cli/template/extras/config/_prettier.config.cjs

* need to test this but.

* removed uneeded resolve

* docs update

* Moved over to a seperate prompt

* yuh

* spelling

* this?

* grammar

* grammer nerd @ryanagillie <3

* spelling

---------

+++ cli/src/index.ts
@@ -32,7 +32,7 @@ const main = async () => {
   const {
     appName,
     packages,
-    flags: { noGit, noInstall, importAlias },
+    flags: { noGit, noInstall, importAlias, prettierAndExtendedEslint },
   } = await runCli();
 
   const usePackages = buildPkgInstallerMap(packages);
@@ -45,6 +45,7 @@ const main = async () => {
     packages: usePackages,
     importAlias: importAlias,
     noInstall,
+    prettierAndExtendedEslint,
   });
 
   // Write name to package.json
@@ -67,10 +68,12 @@ const main = async () => {
   }
 
   // Rename _eslintrc.json to .eslintrc.json - we use _eslintrc.json to avoid conflicts with the monorepos linter
-  fs.renameSync(
-    path.join(projectDir, "_eslintrc.cjs"),
-    path.join(projectDir, ".eslintrc.cjs"),
-  );
+  if (fs.existsSync(path.join(projectDir, "_eslintrc.cjs"))) {
+    fs.renameSync(
+      path.join(projectDir, "_eslintrc.cjs"),
+      path.join(projectDir, ".eslintrc.cjs"),
+    );
+  }
 
   if (!noGit) {
     await initializeGit(projectDir);
